Update hotspot settings subtitle behavior on connected clients changed
When bridged AP is started, there are actually two SoftAP instances.
Currently, the number of connected devices showed in hotspot
settings subtitle is the clients size of latest updated instance.
That is incorrect.

This commit adds up the clients size of both instances to make sure
the number of connected devices in hotspot settings subtitle is correct.

Change-Id: I3351f59d7a9d01d16c4e8b1454144adbc08cd932