Add uiMode to ProfileSwitcherActivity and SubSettingsActivity
When tapping "Add a profile", a new user must be created and switched to.
However, changing Day/Night mode on the user switching screen blocks the switching to a new user because ProfileSwitcherActivity and SubSettingsActivity are generated again with onConfigChanged.
This patch add "uiMode" to android:configChanges in these Activities.

Bug: 319171570
Test: Manual
Change-Id: I6e841ef2d5b54d901668f1f4c86f277e35a1f028