Revert "Fix DoubleBraceInitialization errorprone issues"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/56c3b8e6-d085-4c2e-a5aa-14bc7941815d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I22aab2980c7273a85ae3f465c5551d7a02e4a52f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com