Make InitialLockSetupService accept binary 1-9 in pattern byte[]
The car setup code was using a convention of converting between byte[]
and lists of pattern cells that did not match that used by the
platform's LockPatternUtils.  Commit d8ca2f89b240 ("Simplify lockscreen
credential validation in InitialLockSetupService") unintentionally fixed
this on the service side, which broke compatibility with the client
side.  For now, make the service accept byte[] in both conventions.

Bug: 303716603
Test: mmm packages/apps/Car/Settings
Change-Id: I1ee3460633e7e6a7bf38d393c46a6e28a4c7fdbb