Revert "Do not keep runtime invisible annotations in proguard.flags"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7777bce3-aab5-44d6-ba1f-37966ec29716). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I54ffb66d395aa0c7867d73bcd4c7604c18d05e72
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com