Revert "Do not keep runtime invisible annotations in proguard.flags"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bab71c43-a4e2-44d8-919b-df00081b1810). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I610293ab8e8d2750d58201f1368b82ff71c5014e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com