Revert "Do not keep runtime invisible annotations in proguard.flags"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b2578539-413f-4b43-be59-1252ac64385c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9ff64aa15fcfb8edb581ec65524a4a3177c5b4c5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com