Revert "Do not keep runtime invisible annotations in proguard.flags"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1205eff2-59a5-48f9-9755-aed90664cf32). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id369b578cc4b1958b2d916ee085d8fce1d12ce4b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com