Revert "Do not keep runtime invisible annotations in proguard.flags"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0faad0d4-fade-4a85-9f5e-50ad0a05e0ac). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iaa58c729629b72c6acd28f98ce6825200692bcd3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com