Revert "Do not keep runtime invisible annotations in proguard.flags"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/14b886ad-5ac9-44c7-859d-12db8cca3132). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iaec75bdf473c93c8021c28ed3436c66fa60d5844
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com