Revert "Do not keep runtime invisible annotations in proguard.flags"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d7c35732-3a27-4754-9b35-bd5a5e9fe819). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I59d9844c61c8876d0f4537b50f82afa2893dbdcb
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com