Revert "Do not keep runtime invisible annotations in proguard.flags"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a22f695c-4d60-4de6-86c5-6a3dfaa7abae). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icfbc9fba05beb7a0057df575b8110b668761b796
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com